+++ packages/limm-utools/types/interface/index.d.ts
@@ -1,0 +1,6 @@
+export interface BaseResult<T> {
+  code: Number
+  success: Boolean
+  msg: String
+  data?: T
+}