+++ src/uni_modules/limm-utools/types/core/uni-app/index.d.ts
@@ -5,10 +5,6 @@
  * @LastEditors: SunSeekerX
  * @LastEditTime: 2021-09-17 20:40:42
  */
-/**
- * uni-app 请求类
- */
-import { Request } from './request';
-import { LogUtil, Log } from './log';
 export * from './uni-app';
-export { Request, LogUtil, Log };
+export * from './request';
+export * from './log';